chore: set up unit tests for validation
