+++ src/checker.ts
@@ -32,25 +32,9 @@ export type EventBusMapInterface = ts.CallExpression & {
   arguments: [TsFunctionParameter];
 };
 
-export type FunctionlessChecker = ReturnType<typeof _makeFunctionlessChecker>;
-
-// cache a mapping of checker to functionless checker to avoid redundant computations
-// weak cache because we want to clear memory whenever it removes
-const weakCache = new WeakMap<
-  ts.TypeChecker | tsserver.TypeChecker,
-  FunctionlessChecker
->();
+export type FunctionlessChecker = ReturnType<typeof makeFunctionlessChecker>;
 
 export function makeFunctionlessChecker(
-  checker: ts.TypeChecker | tsserver.TypeChecker
-): FunctionlessChecker {
-  if (!weakCache.has(checker)) {
-    weakCache.set(checker, _makeFunctionlessChecker(checker));
-  }
-  return weakCache.get(checker)!;
-}
-
-function _makeFunctionlessChecker(
   checker: ts.TypeChecker | tsserver.TypeChecker
 ) {
   return {
