+++ test/validate.test.ts
@@ -1,0 +1,41 @@
+import "jest";
+import fs from "fs";
+import path from "path";
+import ts from "typescript";
+import { makeFunctionlessChecker } from "../src/checker";
+import { formatDiagnosticsWithColorAndContext } from "../src/format-error";
+import { validate } from "../src/validate";
+
+const testFilesDir = path.resolve(path.join(__dirname, "test-files"));
+
+const tsconfig = {};
+const compilerHost = ts.createCompilerHost(tsconfig, true);
+
+const fileNames = fs
+  .readdirSync(testFilesDir)
+  .map((fileName) => path.join(testFilesDir, fileName));
+
+const program = ts.createProgram(fileNames, tsconfig, compilerHost);
+
+const checker = makeFunctionlessChecker(program.getTypeChecker());
+
+for (const fileName of fileNames) {
+  test(path.basename(fileName), () => {
+    const diagnostics = validate(ts, checker, program.getSourceFile(fileName)!);
+    const errors = normalize(
+      formatDiagnosticsWithColorAndContext(diagnostics, compilerHost)
+    );
+    expect(errors).toMatchSnapshot();
+  });
+}
+
+function normalize(str: string) {
+  return str.replace(
+    new RegExp(escapeRegExp(testFilesDir), "g"),
+    "<workspace>"
+  );
+}
+
+function escapeRegExp(string: string) {
+  return string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"); // $& means the whole matched string
+}
